several fixes
