+++ requirements.txt
@@ -1,3 +1,3 @@
-bs4
+beautifulsoup4
 requests
 enum34
